+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Added GetAggregatedAttestationV2 endpoint.
 - Added SubmitAttestationsV2 endpoint.
 - Validator REST mode Electra block support
+- Added validator index label to `validator_statuses` metric
 
 ### Changed